Extract generic carousel logic from scrollable small
